fix windows compilation
