+++ docker/sock_windows.go
@@ -1,0 +1,8 @@
+package docker
+
+import "net/url"
+
+// shouldConnectSock is always false on windows
+func shouldConnectSock(host *url.URL) bool {
+	return false
+}
